Add support for SDK name suffix
We want to suffix the SDK-Name HTTP header/query parameter when the SDK is wrapped/embedded in another SDK.

Add new public API property to `OpenPassSettings` to allow setting an SDK name suffix.

I also took the opportunity to move some of the private Client configuration/properties out of `OpenPassManager` to `OpenPassClient`.